Functional Pipeline
The pipeline functions fully, and renames the input files although there may still be some work to do here including
moving this renaming step to before the begining of the pipeline and possibly added as step which removes '-' from the
output names.  Replace with '_'.  This would make transitioning the pipeline to R easier.

[ticket: X]