Runs DeepSNV as bpipe stage
The pipeline has been reworked so that the deepSNV variant caller now runs as part of the bpipe pipeline.  The pipeline is also implemented in a more straightforeward manner.  All we need to do now is add give the bash script the input and output directories the fasta reference and the plasmid control sample name.  All needed varaibles and directories are then written to the .conig file. pretty slick.

[ticket: X]