Updating to handel odd cases
I noticed there was an insertion in one of the samples Vic_pool relative to perth. The insertion was in the noncoding region of NA 3' end - muscle removes this. There was also a 'mutation' on the last base all of this ment the size of the segment was 1 base shorted than the mutation position. if this happens now we call the mutation non-coding and move on.